# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Eventia Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Invitation {
  id: number;
  coupleName1: string;
  coupleName2: string;
  collection: string;
  formula: string;
  blocks: unknown[];
  status: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  eventDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InvitationInput {
  coupleName1: string;
  coupleName2: string;
  collection: string;
  formula: string;
  /** @nullable */
  eventDate?: string | null;
}

export interface InvitationUpdate {
  coupleName1?: string;
  coupleName2?: string;
  collection?: string;
  formula?: string;
  blocks?: unknown[];
  status?: string;
  /** @nullable */
  eventDate?: string | null;
}

export interface Collection {
  id: number;
  slug: string;
  name: string;
  description?: string;
  videoUrl: string;
  previewImageUrl?: string;
}