# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Eventia Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  slug: string;
  name: string;
  description?: string;
  videoUrl: string;
  previewImageUrl?: string;
}