# lib/api-zod/src/generated/types/invitationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Eventia Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface InvitationInput {
  coupleName1: string;
  coupleName2: string;
  collection: string;
  formula: string;
  /** @nullable */
  eventDate?: string | null;
}