# lib/api-zod/src/generated/types/invitationUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Eventia Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface InvitationUpdate {
  coupleName1?: string;
  coupleName2?: string;
  collection?: string;
  formula?: string;
  blocks?: unknown[];
  status?: string;
  /** @nullable */
  eventDate?: string | null;
}